pyproject.toml
[project]
name = "medical-dictation-cli"
version = "0.1.0"
description = "CLI tool for transcribing German medical dictations and extracting structured clinical summaries"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "pydub>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
medical-cli = "medical_dictation.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/medical_dictation"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

